# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: TradeTies API
  version: 0.1.0
  description: |
    Two-sided marketplace connecting customers with local tradespeople.
  contact:
    name: TradeTies
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.tradeties.com
    description: Production
...
security:
  - workosAuthKit: []
paths:
...
  /api/v1/jobs:
    post:
      tags: [job]
      operationId: createJob
      security: []
      summary: Send a customer's first request for a slot.
      description: |
        The first write on the customer side, and the first operation in this API
        that takes no token at all. Everything before it answered a question; this
        one creates something a tradesperson has to act on.

        **It creates two things: the job and its first request.** The job is what
        the customer wants done — the address, the description, how to reach them.
        The request is that job put to one business, for one service, at one start.
        They are separate because DECISIONS section 1 sends the same job to several
        tradespeople in parallel, each deciding on their own; the second and later
        requests are a different operation against a job that already exists.

        **What binds is what was shown.** The caller sends no prices, no duration
        and no cancellation fee. Every one of those is resolved here, from the
        business and the service named, and written onto the request as a snapshot.
        A business that raises its cancellation fee tomorrow must not make today's
        request more expensive, and a client that could send the numbers could send
        different ones.

        **`startsAt` is checked against the same walk that offered it.** It has to
        be a start this business actually offers for this service today — on the
        grid, inside the working week, past the notice, inside the horizon, clear
        of declared time off and of appointments already accepted. A start that
        stopped being offered between the page loading and the form being sent is
        refused here rather than accepted and declined later.

        **Nothing is reserved by this.** The request is `PENDING`, and a pending
        request does not take the slot off anybody else's screen — two customers
        asking for one time is correct behaviour, settled when one of them is
        accepted and not before.

        The answer carries an access token **once, in the clear**. A customer has
        no account, so that token is the only way back to what they sent. It is
        never returned again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: The created job, its first request, and the access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/by-token:
    get:
      tags: [job]
      operationId: getJobByToken
      security: []
      summary: What a customer sent, read back with their access token.
      description: |
        The customer's page, and the other half of the token `createJob` hands out.
        A customer has no account, so this token is how they come back — from the
        link in their confirmation email — to see what they asked for and where it
        stands.

        **The token travels in a header, never in the path.** A path is written into
        access logs, proxy logs and browser history on the way through; a header is
        not. The page the customer opens carries it in its own address, and hands it
        on here without repeating it.

        **An unknown token and an expired one are one answer.** Both are `404`.
        Telling them apart would confirm that a guessed value had once been real.

        The answer is the job and every request on it, each with the terms that were
        frozen when it was sent. Nothing a tradesperson wrote is here yet; messages
        join this page when they exist.
      parameters:
        - name: X-Job-Token
          in: header
          required: true
          description: The access token from `CreatedJob.accessToken`, exactly as issued.
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        "200":
          description: The job and its requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/by-token/requests/{requestId}/messages:
    post:
      tags: [job]
      operationId: sendCustomerMessage
      security: []
      summary: A customer writes to the business their request went to.
      description: |
        The customer's side of the conversation, from the page their link opens.
        The same token as `getJobByToken`, in the same header; the request has to
        belong to the job that token opens, or the answer is `404` — a token for one
        job reaches no other job's conversation.

        The business is told by email that a message arrived.
      parameters:
        - name: X-Job-Token
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: The message, as stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/conversations:
    get:
      tags: [inbox]
      operationId: listMyConversations
      summary: Every conversation with a customer, latest activity first.
      description: |
        One per request sent to the caller's business, whatever its status — a
        declined request can still have something left to say. Each carries its
        latest line and how many of the customer's messages are unread, which is all
        a list needs; the thread itself is `getMyConversation`.

        The request's own description counts as its first line, so a conversation
        with no messages yet still has something to preview.
      responses:
        "200":
          description: The conversations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/conversations/{requestId}:
    get:
      tags: [inbox]
      operationId: getMyConversation
      summary: One conversation, with the request it is about.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The request, the customer, and every message in order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/conversations/{requestId}/messages:
    post:
      tags: [inbox]
      operationId: sendMyMessage
      summary: Reply to the customer.
      description: |
        Stored, and sent to the customer by email with a fresh link back to their
        request. The link is new for every email: the one from their confirmation
        cannot be written again, because only its digest is kept.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: The message, as stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/conversations/{requestId}/read:
    post:
      tags: [inbox]
      operationId: markMyConversationRead
      summary: Everything the customer wrote so far has been seen.
      description: |
        A POST rather than a side effect of reading, so opening a conversation to
        glance at it and marking it read stay two decisions the client makes.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Marked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    workosAuthKit:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: WorkOS AuthKit access token, validated against the WorkOS JWKS endpoint.
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    BadRequest:
      description: The payload violates the schema or a field constraint.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: |
        The addressed resource does not exist **for this caller**. Owner endpoints
        query by owner rather than checking ownership afterwards, so someone else's
        resource is indistinguishable from a missing one — deliberately. A 403 would
        confirm that the thing exists.

        The anonymous side says the same thing about a different rule: a profile that is
        not published is not the customer's to see, so a draft, a suspension and a slug
        nobody holds are one answer. Distinguishing them would publish a moderation
        decision to anybody who can type a URL.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    JobInput:
      type: object
      description: |
        Everything a customer supplies to send their first request: who they are,
        where the work is, what is wrong, and which slot of whose they want.

        **No prices, no duration, no cancellation fee.** Those are resolved on the
        server from `slug` and `serviceId` and written onto the request as a
        snapshot. A client that could send them could send different ones, and what
        binds has to be what was displayed rather than what was posted back.
      required:
        - slug
        - serviceId
        - startsAt
        - customerName
        - customerEmail
        - description
        - street1
        - city
        - state
        - postalCode
      properties:
        slug:
          $ref: "#/components/schemas/Slug"
        serviceId:
          type: string
          format: uuid
          description: |
            Which of that business's services, by the `id` its profile returned. One
            it does not offer, or has since deactivated, answers `400` — the same
            reading the availability operation gives it.
        startsAt:
          type: string
          format: date-time
          description: |
            The start the customer picked, as the instant the availability operation
            returned. Checked against that same walk before anything is written, so
            a start that has stopped being offered is refused here rather than
            written and declined days later.
          examples: ["2026-10-01T14:00:00Z"]
        customerName:
          type: string
          maxLength: 200
          minLength: 1
        customerEmail:
          type: string
          format: email
          maxLength: 320
          description: |
            The only way back to this job for somebody with no account, together
            with the access token the answer carries.
        customerPhone:
          type: string
          maxLength: 16
          description: Optional. Some jobs are settled faster by phone than by mail.
        description:
          type: string
          minLength: 1
          maxLength: 2000
          description: |
            What is wrong, in the customer's own words. Not the service name — that
            is already known from `serviceId`, and this is the part the tradesperson
            reads to decide whether the estimate still holds.
        street1:
          type: string
          maxLength: 200
          minLength: 1
        street2:
          type: string
          maxLength: 200
          description: Optional. Flat, floor, entrance.
        city:
          type: string
          maxLength: 100
          minLength: 1
        state:
          $ref: "#/components/schemas/UsStateCode"
        postalCode:
          type: string
          pattern: '^\d{5}(-\d{4})?$'
          maxLength: 10
          description: |
            The job site, which is not necessarily where the customer is. Only the
            five digits are read, and they are what places the site on the map.
    CreatedJob:
      type: object
      description: |
        What was written, and the one credential that comes with it.
      required: [jobId, accessToken, accessTokenExpiresAt, request]
      properties:
        jobId:
          type: string
          format: uuid
        accessToken:
          type: string
          description: |
            **Returned once, in the clear, and never again.** Only its SHA-256 is
            stored, so this response is the only moment it exists in readable form.
            A customer has no account and no session; this token is the whole of
            their way back to what they sent.

            Treat it as a credential. It belongs in a link the customer keeps, not
            in a page anybody else can reach.
        accessTokenExpiresAt:
          type: string
          format: date-time
        request:
          $ref: "#/components/schemas/JobRequestSummary"
    CustomerJob:
      type: object
      description: |
        A job as its customer sees it: what they told us, where, and every request
        they sent about it. Their own details only — nothing here belongs to anybody
        who does not already hold the token.
      required:
        - jobId
        - customerName
        - customerEmail
        - description
        - address
        - accessTokenExpiresAt
        - requests
      properties:
        jobId:
          type: string
          format: uuid
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          description: Absent when they gave none.
        description:
          type: string
          description: What they need done, in their own words.
        address:
          $ref: "#/components/schemas/Address"
        accessTokenExpiresAt:
          type: string
          format: date-time
          description: |
            When this page stops opening. Said on the page, so a customer is not
            surprised by a link that worked last month.
        requests:
          type: array
          description: Newest first, each with its conversation.
          items:
            $ref: "#/components/schemas/CustomerRequest"
    CustomerRequest:
      type: object
      description: One request as its customer sees it, and the conversation about it.
      required: [request, messages]
      properties:
        request:
          $ref: "#/components/schemas/JobRequestSummary"
        messages:
          type: array
          description: Oldest first.
          items:
            $ref: "#/components/schemas/Message"
    Message:
      type: object
      required: [id, author, body, sentAt]
      properties:
        id:
          type: string
          format: uuid
        author:
          $ref: "#/components/schemas/MessageAuthor"
        body:
          type: string
        sentAt:
          type: string
          format: date-time
    MessageAuthor:
      type: string
      description: |
        Which side wrote it. The business rather than a person: the customer is
        talking to the business they chose, whoever on its side answers.
      enum:
        - BUSINESS
        - CUSTOMER
    MessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 5000
    ConversationSummary:
      type: object
      description: One line of the inbox.
      required: [requestId, customerName, request, lastMessage, unreadCount]
      properties:
        requestId:
          type: string
          format: uuid
        customerName:
          type: string
        request:
          $ref: "#/components/schemas/JobRequestSummary"
        lastMessage:
          $ref: "#/components/schemas/Message"
          description: |
            The newest line — the customer's description when nobody has written
            since, with the request's id as its id and the request's time as its
            time.
        unreadCount:
          type: integer
          format: int32
    Conversation:
      type: object
      description: |
        Everything the tradesperson needs to answer: the request, who sent it, what
        they wrote, and every message since.
      required:
        - request
        - customerName
        - customerEmail
        - description
        - address
        - requestedAt
        - messages
      properties:
        request:
          $ref: "#/components/schemas/JobRequestSummary"
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        description:
          type: string
          description: What the customer wrote when they sent the request.
        address:
          $ref: "#/components/schemas/Address"
        requestedAt:
          type: string
          format: date-time
          description: When the request was sent, the moment the description was written.
        messages:
          type: array
          description: Oldest first.
          items:
            $ref: "#/components/schemas/Message"
    JobRequestSummary:
      type: object
      description: |
        One request as it now stands, including the terms snapshotted onto it.
        Echoed back rather than left to the client's memory: these are the numbers
        that bind, and the answer is where a customer sees what was recorded.
      required:
        - id
        - status
        - businessSlug
        - businessName
        - serviceName
        - estimatedDurationMinutes
        - startsAt
        - endsAt
        - timeZone
        - currency
        - cancellationFee
        - cancellationNoticeHours
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/JobRequestStatus"
        businessSlug:
          $ref: "#/components/schemas/Slug"
        businessName:
          type: string
          maxLength: 200
        serviceName:
          type: string
          maxLength: 160
          description: |
            The name as it read when the request was sent. A business that renames
            the service afterwards does not rename what somebody asked for.
        estimatedDurationMinutes:
          type: integer
          format: int32
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
          description: |
            `startsAt` plus the duration above, computed here rather than sent. It is
            what the overlap check between accepted appointments is written against,
            so it is not a number a client gets to propose.
        timeZone:
          type: string
          description: |
            The zone this appointment is read against, recorded on the job.

            It is the business's own. DECISIONS section 1 calls for the job site's,
            and the two are one thing here until there is data to tell them apart:
            the postal code table carries coordinates and no zone, so a site's own
            zone cannot be derived today. Recording the tradesperson's is the honest
            reading — the slot came out of their working week.
          examples: ["America/Denver"]
        currency:
          type: string
        cancellationFee:
          $ref: "#/components/schemas/MoneyAmount"
        cancellationNoticeHours:
          type: integer
          format: int32
          description: |
            With the fee above, the pair a customer had to have read before sending.
            Both are frozen here exactly as they were displayed.
    JobRequestStatus:
      type: string
      description: |
        Where one request stands. Only `PENDING` can occur today — nothing can
        accept, decline or withdraw yet.

        **All six are declared now anyway.** A new value in a response enum reaches
        clients that may not handle it: Jackson rejects unknown enum values by
        default, and an exhaustive TypeScript switch falls through in silence.
        Adding them later would be a breaking change bought for a shorter list
        today.
      enum:
        - PENDING
        - ACCEPTED
        - DECLINED
        - WITHDRAWN
        - CANCELLED
        - COMPLETED
...
    UsStateCode:
      type: string
      description: Two-letter USPS state code, upper case. Includes `DC`.
      pattern: '^[A-Z]{2}$'
      minLength: 2
      maxLength: 2
      examples: ["CO"]
...
    Slug:
      type: string
      description: |
        The public URL segment: `tradeties.com/{slug}`. Lower case letters,
        digits and single hyphens. Proposed from the display name in the wizard,
        then checked for availability.

        **Fixed from the first publish onwards.** While the profile is a draft it is
        not searchable and no link to it exists, so it can still be changed freely.
        Publishing turns it into an address that gets printed on a van and read out
        over the phone, and nothing this API can do would repair those copies. From
        then on a differing `slug` on `PUT /api/v1/me/business` is a 409, and
        `BusinessProfile.slugLocked` says so before the attempt is made.

        The display name proposes it and does not own it: the two part company the
        moment a business is renamed, and it is the URL that stays.

        **Some slugs are reserved.** Profiles share the root of the domain with the
        site's own pages, so names like `portal`, `help` or `pricing` cannot be
        taken: creating with one, or moving to one, is a 409. A business that held a
        name before it was reserved keeps it.
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      maxLength: 80
      examples: ["joes-plumbing"]
    Address:
      type: object
      description: |
        A US postal address. Its own schema rather than inline properties because
        the job site will use the same shape.
      required: [street1, city, state, postalCode]
      properties:
        street1:
          type: string
          minLength: 1
          maxLength: 200
          examples: ["123 Main St"]
        street2:
          type: [string, "null"]
          maxLength: 200
          examples: ["Suite 4"]
        city:
          type: string
          minLength: 1
          maxLength: 100
          examples: ["Denver"]
        state:
          $ref: "#/components/schemas/UsStateCode"
        postalCode:
          type: string
          description: ZIP or ZIP+4.
          pattern: '^\d{5}(-\d{4})?$'
          maxLength: 10
          examples: ["80202"]
...
    Problem:
      type: object
      description: |
        RFC 9457 problem detail, as produced by Spring's ProblemDetail.

        `type` is `about:blank` unless a problem is one a client has to tell apart from the
        others sharing its status code. One is a `400` and the rest are `409`s — the two
        statuses this API answers with for more than one reason:

        - `urn:tradeties:problem:invalid-selection` — a `400`. The request is well formed, but
          a value it names is not in a catalogue the server keeps: a state, a trade, a time
          zone, a postal code. Name a different one. The other `400` — a request outside the
          bounds the contract itself declares — carries no type, because nobody chose it and
          the person in front of the client cannot fix it.
        - `urn:tradeties:problem:slot-not-offered` — a `400` on sending a request. The start
          named is not one that business is offering for that service any more: somebody
          else's request was accepted onto it, time off was entered over it, or the page has
          simply been open a while. Read the calendar again and pick another time — this is
          not the same remedy as `invalid-selection`, which sends the customer back to the
          service list instead.
        - `urn:tradeties:problem:unconfirmed-time-zone-change` — answer with a dialogue and
          send the same body again with `timeZoneChangeConfirmed`.
        - `urn:tradeties:problem:slug-locked` — the profile was published, so `slug` is no
          longer the holder's to change. A client that believed otherwise is holding a stale
          `slugLocked`: re-read the profile and offer the stored URL, rather than resending.
        - `urn:tradeties:problem:unconfirmed-unpublish` — removing that service would take a
          published profile off the marketplace. Put `detail` to the holder and send the same
          request again with `unpublishConfirmed`, or send nothing.
        - `urn:tradeties:problem:live-profile-not-ready` — the change would leave a published
          profile failing a condition it was meeting, so it was refused and rolled back.
          Nothing to confirm and nothing to re-read: the version you sent still stands. Show
          `detail` and let the holder undo the edit.

        Match on `type`, never on `detail` — the wording is written for people and will be
        reworded.

        A `409` with no `type` on an operation that takes a `version` means the version sent
        was not the stored one. That is why the ones above carry a type: without it a client
        cannot tell a stale version from a refusal it has to show.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri